Split up the avx logic; files gettin' big
