+++ src/avx/mod.rs
@@ -1,0 +1,6 @@
+mod labs_to_rgbs;
+mod rgbs_to_labs;
+
+use super::Lab;
+pub use self::labs_to_rgbs::*;
+pub use self::rgbs_to_labs::*;
